Take steps towards working when called in Electron
The reason that Nightmare doesn't work when called from within Electron
is that it uses `require('electron')` to grab the `electron-prebuilt`
location but `require('electron')` returns the Electron APIs when called
within an Electron app. So I wrote `electron-prebuilt-path`, which
specifically grabs the `electron` package provided by
`electron-prebuilt` even if used within Electron.